@@ -16,7 +16,7 @@
         }
     },
     "require": {
-        "php": "^7.2",
+        "php": "^7.4 || ^8.0",
         "ext-json": "*",
         "nyholm/psr7": "^1.2",
         "php-http/httplug": "^2.0",
@@ -25,14 +25,17 @@
         "thecodingmachine/safe": "^1.0.2"
     },
     "require-dev": {
-        "doctrine/coding-standard": "^6.0",
+        "cdn77/coding-standard": "^4.0",
+        "hectorj/safe-php-psalm-plugin": "dev-master#b60ed45a06d5246e2efd193ce9c8940b244d5c7d",
         "jakub-onderka/php-parallel-lint": "^1.0",
         "phpstan/extension-installer": "^1.0",
-        "phpstan/phpstan": "^0.11.16",
-        "phpstan/phpstan-phpunit": "^0.11.2",
-        "phpstan/phpstan-strict-rules": "^0.11.1",
-        "phpunit/phpunit": "^8.4",
-        "thecodingmachine/phpstan-safe-rule": "^1.0"
+        "phpstan/phpstan": "0.12.81",
+        "phpstan/phpstan-phpunit": "0.12.18",
+        "phpstan/phpstan-strict-rules": "^0.12.7",
+        "phpunit/phpunit": "^9.5",
+        "psalm/plugin-phpunit": "0.15.1",
+        "thecodingmachine/phpstan-safe-rule": "^1.0",
+        "vimeo/psalm": "4.6.3"
     },
     "config": {
         "sort-packages": true
