@@ -8,7 +8,7 @@ use Psr\Http\Message\RequestInterface;
 
 interface JsonRpcRequestFactory
 {
-    public const METHOD_REQUEST      = 'REQUEST';
+    public const METHOD_REQUEST = 'REQUEST';
     public const METHOD_NOTIFICATION = 'NOTIFICATION';
 
     /**
